+++ lib/live_view_native/jetpack/client.ex
@@ -1,0 +1,8 @@
+defmodule LiveViewNative.Jetpack.Client do
+  @moduledoc false
+
+  defstruct tags: %{
+    form: "LiveForm",
+    button: "Button"
+  }
+end