+++ test/live_view_native/jetpack/integrations/inline_test.exs
@@ -2,7 +2,6 @@ defmodule LiveViewNative.Jetpack.InlineRenderTest do
   use ExUnit.Case, async: false
 
   import Phoenix.ConnTest
-  import Phoenix.LiveViewTest
   import LiveViewNativeTest
 
   @endpoint LiveViewNativeTest.Endpoint
@@ -12,14 +11,14 @@ defmodule LiveViewNative.Jetpack.InlineRenderTest do
   end
 
   test "can render the jetpack format", %{conn: conn} do
-    {:ok, lv, _body} = native(conn, "/inline", :jetpack)
+    {:ok, lv, _body} = live(conn, "/inline", _format: :jetpack)
 
-    assert lv |> element("text") |> render() =~ "Inline Jetpack Render 100"
+    assert lv |> element("Text") |> render() =~ "Inline Jetpack Render 100"
   end
 
   test "can render the jetpack format with watch target", %{conn: conn} do
-    {:ok, lv, _body} = native(conn, "/inline", :jetpack, %{"target" => "watch"})
+    {:ok, lv, _body} = live(conn, "/inline", _format: :jetpack, _interface: %{"target" => "watch"})
 
-    assert lv |> element("text") |> render() =~ "Watch Target Inline Jetpack Render 100"
+    assert lv |> element("Text") |> render() =~ "Watch Target Inline Jetpack Render 100"
   end
 end