+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LiveViewNative.Jetpack.normalize_os_version/1`
 - `LiveViewNative.Jetpack.normalize_app_version/1`
 - CoreComponents
+- LiveViewNative.Jetpack.Client
 
 ## Changed